requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
twilio==8.11.0
pydantic==2.5.3
python-dotenv==1.0.0